Rename SAPDBDialect to MaxDBDialect
(Leaving a stub SAPDBDialect class for backward compatibility.)
