+++ etc/hibernate.properties.template
@@ -119,9 +119,9 @@ hibernate.connection.url @DB_URL@
 #hibernate.connection.url jdbc:sybase:Tds:co3061835-a:5000/tempdb
 
 
-## SAP DB
+## SAP MaxDB
 
-#hibernate.dialect org.hibernate.dialect.SAPDBDialect
+#hibernate.dialect org.hibernate.dialect.MaxDBDialect
 #hibernate.connection.driver_class com.sap.dbtech.jdbc.DriverSapDB
 #hibernate.connection.url jdbc:sapdb://localhost/TST
 #hibernate.connection.username TEST
